Rename layouts folder to layout
